+++ src/views/Settings.vue
@@ -12,6 +12,11 @@
     </v-row>
     <v-row v-if="isStaff">
       <v-col sm="12">
+        <DiagnosticsCard></DiagnosticsCard>
+      </v-col>
+    </v-row>
+    <v-row v-if="isStaff">
+      <v-col sm="12">
         <RetentionCard></RetentionCard>
       </v-col>
     </v-row>
@@ -22,10 +27,11 @@
   import CommonSettings from "../components/CommonSettings"
   import UserSettings from "../components/UserSettings"
   import RetentionCard from "../components/RetentionCard"
+  import DiagnosticsCard from "../components/DiagnosticsCard"
 
   export default {
     name: "Settings",
-    components: { UserSettings, CommonSettings, RetentionCard },
+    components: { UserSettings, CommonSettings, RetentionCard, DiagnosticsCard },
     computed: {
       isStaff() {
         return this.$store.getters.isStaff